remove ground_truth_nsrts
